Refactor a few defines to classes, improve dependencies
Modify maxscale::install, maxscale::config, maxscale::install::apt to be
classes instead of defines.

Remove file resource for real_configdir, it causes a dependency issue
if it is /etc.  This could be moved to the install class if it is
required.

Improve depedency between classes and resources such that install
resources happen before config, and config changes trigger a service
restart.